baselines added

+++ archive/batch1/c1105.c
@@ -1,10 +1,11 @@
 #include <stdio.h>
+#include <stdint.h>
 
-struct Register { union { unsigned word; struct { unsigned low; unsigned high; }; }; };
+struct Register { union { uint16_t word; struct { uint8_t low; uint8_t high; }; }; };
 int main(void)
 {
     struct Register r;
     r.word = 0x1234;
-    printf("c1105 register=%X/%X\n", r.low, r.high);
+    printf("c1105 register=%X/%X\n", (unsigned)r.low, (unsigned)r.high);
     return 0;
 }
