+++ archive/batch2/c1108.c
@@ -1,3 +1,4 @@
 #include <stdio.h>
-struct Register{union{unsigned word;struct{unsigned low;unsigned high;};};};
-int main(void){struct Register r;r.word=0x1234;printf("c1108 register=%X/%X\n",r.low,r.high);return 0;}
+#include <stdint.h>
+struct Register{union{uint16_t word;struct{uint8_t low;uint8_t high;};};};
+int main(void){struct Register r;r.word=0x1234;printf("c1108 register=%X/%X\n",(unsigned)r.low,(unsigned)r.high);return 0;}
