+++ archive/batch2/c8910.c
@@ -1,3 +1,3 @@
 #include <stdio.h>
-static int power(int base,int exponent){int r=1;while(exponent){if(exponent%2)r=r*base%1000;base=base*base%1000;exponent/=2;}return r;}
-int main(void){printf("c8910 modpow=%d\n",power(7,23));return 0;}
+static long power(long base,int exponent){long r=1;while(exponent){if(exponent%2)r=r*base%1000L;base=base*base%1000L;exponent/=2;}return r;}
+int main(void){printf("c8910 modpow=%ld\n",power(7L,23));return 0;}
