+++ archive/batch5/c9929.c
@@ -1,3 +1,4 @@
 #include <stdio.h>
-static int pack(int hi,int lo){return (hi<<8)|lo;}
-int main(void){printf("c9929 pack=%d,%d\n",pack(0x12,0x34),pack(0xAB,0xCD));return 0;}
+#include <stdint.h>
+static uint16_t pack(uint8_t hi,uint8_t lo){return (uint16_t)(((uint16_t)hi<<8)|lo);}
+int main(void){printf("c9929 pack=%u,%u\n",(unsigned)pack(0x12,0x34),(unsigned)pack(0xAB,0xCD));return 0;}
