+++ archive/batch6/c9922.c
@@ -1,3 +1,4 @@
 #include <stdio.h>
-static int encode_pair(int hi,int lo){return (hi<<8)|lo;}
-int main(void){printf("c9922 pair=%d,%d\n",encode_pair(0x12,0x34),encode_pair(0xAB,0xCD));return 0;}
+#include <stdint.h>
+static uint16_t encode_pair(uint8_t hi,uint8_t lo){return (uint16_t)(((uint16_t)hi<<8)|lo);}
+int main(void){printf("c9922 pair=%u,%u\n",(unsigned)encode_pair(0x12,0x34),(unsigned)encode_pair(0xAB,0xCD));return 0;}
